replace cell statistics to single json column
