--- studio/alembic/versions/5a565ec4961f_add_cell_attributes.py
@@ -1,51 +1,0 @@
-"""add cell attributes
-
-Revision ID: 5a565ec4961f
-Revises: 206c32e0128b
-Create Date: 2023-09-12 11:06:03.209355
-
-"""
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision = "5a565ec4961f"
-down_revision = "206c32e0128b"
-branch_labels = None
-depends_on = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("cells", sa.Column("p_value_resp", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("p_value_sel", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("p_value_ori_resp", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("p_value_ori_sel", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("dir_vector_angle", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("ori_vector_angle", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("oi", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("di", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("dsi", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("osi", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("r_best_dir", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("dir_tuning_width", sa.Float(), nullable=True))
-    op.add_column("cells", sa.Column("ori_tuning_width", sa.Float(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("cells", "ori_tuning_width")
-    op.drop_column("cells", "dir_tuning_width")
-    op.drop_column("cells", "r_best_dir")
-    op.drop_column("cells", "osi")
-    op.drop_column("cells", "dsi")
-    op.drop_column("cells", "di")
-    op.drop_column("cells", "oi")
-    op.drop_column("cells", "ori_vector_angle")
-    op.drop_column("cells", "dir_vector_angle")
-    op.drop_column("cells", "p_value_ori_sel")
-    op.drop_column("cells", "p_value_ori_resp")
-    op.drop_column("cells", "p_value_sel")
-    op.drop_column("cells", "p_value_resp")
-    # ### end Alembic commands ###
