+++ studio/alembic/versions/b66427b57c69_add_cell_attributes.py
@@ -1,0 +1,27 @@
+"""add cell attributes
+
+Revision ID: b66427b57c69
+Revises: 206c32e0128b
+Create Date: 2023-09-12 22:54:37.919501
+
+"""
+import sqlalchemy as sa
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision = "b66427b57c69"
+down_revision = "206c32e0128b"
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column("cells", sa.Column("statistics", sa.JSON(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column("cells", "statistics")
+    # ### end Alembic commands ###
