+++ studio/app/optinist/core/expdb/crud_cells.py
@@ -35,19 +35,21 @@ def bulk_insert_cells(
             {
                 "cell_number": i,
                 "experiment_uid": experiment_uid,
-                "p_value_resp": p_value_resp[i],
-                "p_value_sel": p_value_sel[i],
-                "p_value_ori_resp": p_value_ori_resp[i],
-                "p_value_ori_sel": p_value_ori_sel[i],
-                "dir_vector_angle": dir_vector_angle[i],
-                "ori_vector_angle": ori_vector_angle[i],
-                "oi": oi[i],
-                "di": di[i],
-                "dsi": dsi[i],
-                "osi": osi[i],
-                "r_best_dir": r_best_dir[i],
-                "dir_tuning_width": dir_tuning_width[i],
-                "ori_tuning_width": ori_tuning_width[i],
+                "statistics": {
+                    "p_value_resp": p_value_resp[i],
+                    "p_value_sel": p_value_sel[i],
+                    "p_value_ori_resp": p_value_ori_resp[i],
+                    "p_value_ori_sel": p_value_ori_sel[i],
+                    "dir_vector_angle": dir_vector_angle[i],
+                    "ori_vector_angle": ori_vector_angle[i],
+                    "oi": oi[i],
+                    "di": di[i],
+                    "dsi": dsi[i],
+                    "osi": osi[i],
+                    "r_best_dir": r_best_dir[i],
+                    "dir_tuning_width": dir_tuning_width[i],
+                    "ori_tuning_width": ori_tuning_width[i],
+                },
             }
             for i in range(stat_data.ncells)
         ],
